services/integration_gateway/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
PyJWT==2.9.0
sqlalchemy==2.0.36
asyncpg==0.29.0
prometheus-fastapi-instrumentator==7.0.0
